Move components into seperate folders
