+++ plugins/user-settings/src/components/AuthProviders/DefaultProviderSettings.test.tsx
