+++ plugins/user-settings/src/components/AuthProviders/DefaultProviderSettings.tsx
